extract parsing entry from args into own method
will be reused
